+++ src/main/kotlin/io/unthrottled/doki/config/ThemeConfig.kt
@@ -45,6 +45,14 @@ class ThemeConfig : PersistentStateComponent<ThemeConfig>, Cloneable {
   var isMaterialFiles: Boolean = false
   var isMaterialPSIIcons: Boolean = false
 
+  var capStickerDimensions: Boolean = false
+  var maxStickerWidth: Int = -1
+  var maxStickerHeight: Int = -1
+
+  var showSmallStickers: Boolean = false
+  var smallMaxStickerWidth: Int = 100
+  var smallMaxStickerHeight: Int = 100
+
   var isSeeThroughNotifications: Boolean = false
   var notificationOpacity: Int = 90