+++ src/main/kotlin/io/unthrottled/doki/settings/ThemeSettings.kt
@@ -51,6 +51,12 @@ data class ThemeSettingsModel(
   var isCustomFontSize: Boolean,
   var customFontSizeValue: Int,
   var isOverrideConsoleFont: Boolean,
+  var capStickerDimensions: Boolean,
+  var maxStickerHeight: Int,
+  var maxStickerWidth: Int,
+  var showSmallStickers: Boolean,
+  var smallMaxStickerHeight: Int,
+  var smallMaxStickerWidth: Int,
   var consoleFontValue: String,
   var isSeeThroughNotifications: Boolean,
   var notificationOpacity: Int,
@@ -95,6 +101,12 @@ object ThemeSettings {
       notificationOpacity = ThemeConfig.instance.notificationOpacity,
       isOverrideConsoleFont = ThemeConfig.instance.isOverrideConsoleFont,
       consoleFontValue = ThemeConfig.instance.consoleFontName,
+      maxStickerHeight = ThemeConfig.instance.maxStickerHeight,
+      maxStickerWidth = ThemeConfig.instance.maxStickerWidth,
+      capStickerDimensions = ThemeConfig.instance.capStickerDimensions,
+      smallMaxStickerHeight = ThemeConfig.instance.smallMaxStickerHeight,
+      smallMaxStickerWidth = ThemeConfig.instance.smallMaxStickerWidth,
+      showSmallStickers = ThemeConfig.instance.showSmallStickers,
     )
 
   fun apply(themeSettingsModel: ThemeSettingsModel) {
@@ -106,6 +118,16 @@ object ThemeSettings {
       themeSettingsModel.customStickerPath,
       themeSettingsModel.isCustomSticker,
       false
+    )
+    StickerActor.setDimensionCapping(
+      themeSettingsModel.capStickerDimensions,
+      themeSettingsModel.maxStickerWidth,
+      themeSettingsModel.maxStickerHeight,
+    )
+    StickerActor.setSmolStickers(
+      themeSettingsModel.showSmallStickers,
+      themeSettingsModel.smallMaxStickerWidth,
+      themeSettingsModel.smallMaxStickerHeight,
     )
     ThemedTitleBarActor.enableThemedTitleBar(themeSettingsModel.isThemedTitleBar)
     ThemeActor.applyTheme(themeSettingsModel.currentTheme)