+++ src/main/kotlin/io/unthrottled/doki/settings/actors/StickerActor.kt
@@ -6,6 +6,7 @@ import io.unthrottled.doki.stickers.CustomStickerService
 import io.unthrottled.doki.stickers.StickerComponent
 import io.unthrottled.doki.stickers.StickerLevel
 import io.unthrottled.doki.stickers.StickerPaneService
+import io.unthrottled.doki.stickers.StickerType
 import io.unthrottled.doki.themes.ThemeManager
 import io.unthrottled.doki.util.performWithAnimation
 
@@ -37,7 +38,7 @@ object StickerActor {
         }
       } else {
         ThemeConfig.instance.stickerLevel = StickerLevel.OFF.name
-        StickerPaneService.instance.remove()
+        StickerPaneService.instance.remove(StickerType.ALL)
       }
     }
   }
@@ -59,10 +60,58 @@ object StickerActor {
 
     val shouldReDraw = isNewStickerPath || isCustomStickersChanged
     if (shouldReDraw) {
-      performWithAnimation(withAnimation) {
-        ThemeManager.instance.currentTheme.ifPresent {
-          StickerPaneService.instance.activateForTheme(it)
-        }
+      activateNewSticker(withAnimation)
+    }
+  }
+
+  fun setDimensionCapping(
+    capStickerDimensions: Boolean,
+    maxStickerWidth: Int,
+    maxStickerHeight: Int
+  ) {
+    val ogWidth = ThemeConfig.instance.maxStickerWidth
+    ThemeConfig.instance.maxStickerWidth = maxStickerWidth
+    val ogHeight = ThemeConfig.instance.maxStickerHeight
+    ThemeConfig.instance.maxStickerHeight = maxStickerHeight
+    val ogActivation = ThemeConfig.instance.capStickerDimensions
+    ThemeConfig.instance.capStickerDimensions = capStickerDimensions
+    if (
+      ogHeight != maxStickerHeight ||
+      ogWidth != maxStickerWidth ||
+      ogActivation != capStickerDimensions
+    ) {
+      activateNewSticker(false)
+    }
+  }
+
+  fun setSmolStickers(
+    isSmolStickers: Boolean,
+    smolMaxStickerWidth: Int,
+    smolMaxStickerHeight: Int
+  ) {
+    val ogWidth = ThemeConfig.instance.smallMaxStickerWidth
+    ThemeConfig.instance.smallMaxStickerWidth = smolMaxStickerWidth
+    val ogHeight = ThemeConfig.instance.smallMaxStickerHeight
+    ThemeConfig.instance.smallMaxStickerHeight = smolMaxStickerHeight
+    val ogActivation = ThemeConfig.instance.showSmallStickers
+    ThemeConfig.instance.showSmallStickers = isSmolStickers
+    if (
+      ogHeight != smolMaxStickerHeight ||
+      ogWidth != smolMaxStickerWidth ||
+      ogActivation != isSmolStickers
+    ) {
+      if (ThemeConfig.instance.showSmallStickers) {
+        activateNewSticker(false)
+      } else {
+        StickerPaneService.instance.remove(StickerType.SMOL)
+      }
+    }
+  }
+
+  private fun activateNewSticker(withAnimation: Boolean) {
+    performWithAnimation(withAnimation) {
+      ThemeManager.instance.currentTheme.ifPresent {
+        StickerPaneService.instance.activateForTheme(it)
       }
     }
   }